ImageBinPlotFOVs

This function generates bin plots for multiple fields of view (FOVs) from a spatial transcriptomics object. If no FOVs are specified, it defaults to using all FOVs available in the object. The function allows customization of bin size, color palette, and scaling options.

Usage

ImageBinPlotFOVs(
  obj,
  feature,
  group.by = NULL,
  fovs = NULL,
  bin_size = 10,
  min = 0,
  max_quantile = "q75",
  x_min = NULL,
  x_max = NULL,
  y_min = NULL,
  y_max = NULL,
  x_col = "x",
  y_col = "y",
  type = "expression",
  assay = "Xenium",
  layer = "count",
  ident_alpha = 0.3,
  ident_pointsize = 1,
  palette = c("viridis", "inferno", "magma", "plasma", "cividis", "mako", "rocket",
    "turbo"),
  palette_begin = 0,
  palette_end = 1,
  filter_ident = NULL,
  feature_on_top = TRUE,
  same_scale = FALSE,
  ncol = NULL
)

Arguments

obj

A spatial transcriptomics object containing the data to be plotted.

feature

The feature to be visualized (e.g., gene expression).

group.by

Optional grouping variable for coloring points.

fovs

A vector of FOVs to be plotted. Defaults to NULL, which uses all FOVs in the object.

bin_size

The size of the bins for aggregating data. Default is 10.

min

Minimum value for the color scale. Default is 0.

max_quantile

Maximum quantile for the color scale. Default is "q75".

x_min,

x_max, y_min, y_max Optional limits for the x and y axes.

x_col,

y_col Column names for x and y coordinates. Default are "x" and "y".

type

The type of data to plot (e.g., "expression"). Default is "expression".

assay

The assay to use for data extraction. Default is "Xenium".

layer

The data layer to use. Default is "count".

ident_alpha

Transparency level for points. Default is 0.3.

ident_pointsize

Size of points in the plot. Default is 1.

palette

A vector of color palettes to choose from. Default includes 'viridis', 'inferno', etc.

palette_begin,

palette_end Range of the color palette to use. Default is 0 to 1.

filter_ident

Optional filter for specific identities.

feature_on_top

Logical, whether to plot the feature layer on top. Default is TRUE.

same_scale

Logical, whether to use the same scale for all FOVs. Default is FALSE.

ncol

Number of columns for arranging plots. Default is NULL.

Value

A combined plot of all FOVs, optionally with the same scale.

See also

Examples

# Generate plots for all FOVs in an object
ImageBinPlotFOVs(obj = my_object, feature = "GeneA")
#> Error in Images(obj): could not find function "Images"

# Generate plots for specific FOVs with custom bin size
ImageBinPlotFOVs(obj = my_object, feature = "GeneA", fovs = c("FOV1", "FOV2"), bin_size = 20)
#> Error in map(fovs, function(fov) {    p <- ImageBinPlot(obj = obj, feature = feature, group.by = group.by,         x_col = x_col, y_col = y_col, fov = fov, x_min = x_min,         x_max = x_max, y_min = y_min, y_max = y_max, bin_size = bin_size,         min = min, max_quantile = max_quantile, type = type,         assay = assay, layer = layer, ident_alpha = ident_alpha,         ident_pointsize = ident_pointsize, filter_ident = filter_ident,         feature_on_top = feature_on_top)    return(p)}): ℹ In index: 1.
#> Caused by error in `Images()`:
#> ! could not find function "Images"

# Generate plots with the same scale across all FOVs
ImageBinPlotFOVs(obj = my_object, feature = "GeneA", same_scale = TRUE)
#> Error in Images(obj): could not find function "Images"